layerscape: reorder/rename ethernet interfaces on Traverse LS1043V
On the LS1043, the layout of the frame manager MACs in memory
is not the same as the physical layout (i.e QSGMII0-1,RGMII0-1,
QSGMII2-3). Reorder the eth interfaces to be the same
as expected by the user.

u-boot already loads the MAC addreses into DT in the physical
order.

Signed-off-by: Mathew McBride <matt@traverse.com.au>